requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
altair==5.0.0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.1.0
apispec==6.3.0
appdirs==1.4.4
APScheduler==3.10.4
argcomplete==3.2.1
asgiref==3.7.2
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
cattrs==23.2.3
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigArgParse==1.7
ConfigUpdater==3.2
connection-pool==0.0.3
connexion==2.14.2
contourpy==1.2.0
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
cycler==0.12.1
datrie==0.8.2
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
dpath==2.1.6
email-validator==1.3.1
et-xmlfile==1.1.0
fastjsonschema==2.19.0
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Script==2.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.46.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-re2==1.1
googleapis-common-protos==1.62.0
graphviz==0.20.1
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.5.0
kiwisolver==1.4.5
kombu==5.3.4
lazy-object-proxy==1.9.0
limits==3.7.0
linkify-it-py==2.0.2
llvmlite==0.41.1
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
nbformat==5.9.2
numba==0.58.1
numpy==1.26.2
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pathspec==0.12.1
pendulum==2.1.2
Pillow==10.1.0
plac==1.4.2
platformdirs==4.1.0
pluggy==1.3.0
prison==0.2.1
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
psycopg2==2.9.9
PuLP==2.7.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.11
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-toolbelt==1.0.0
reretry==0.11.8
rfc3339-validator==0.1.4
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.2
scikit-learn==1.3.2
scipy==1.11.4
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snakemake==7.32.4
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stopit==1.1.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
throttler==1.2.2
toolz==0.12.0
toposort==1.10
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
umap-learn==0.5.5
unicodecsv==0.14.1
urllib3==2.1.0
vine==5.1.0
vl-convert-python==1.2.0
wcwidth==0.2.12
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.0.1
yarl==1.9.4
yte==1.5.4
zipp==3.17.0
